+++ core/common/src/main/res/drawable/ic_arrow_back.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M640,880 L240,480l400,-400 71,71 -329,329 329,329 -71,71Z"
+      android:fillColor="#5f6368"/>
+</vector>
