+++ core/common/src/main/res/drawable/ic_detail_gallary.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="18dp"
+    android:height="19dp"
+    android:viewportWidth="18"
+    android:viewportHeight="19">
+  <path
+      android:pathData="M18,13.1V2.3C18,1.31 17.19,0.5 16.2,0.5H5.4C4.41,0.5 3.6,1.31 3.6,2.3V13.1C3.6,14.09 4.41,14.9 5.4,14.9H16.2C17.19,14.9 18,14.09 18,13.1ZM8.1,9.5L9.927,11.939L12.6,8.6L16.2,13.1H5.4L8.1,9.5ZM0,4.1V16.7C0,17.69 0.81,18.5 1.8,18.5H14.4V16.7H1.8V4.1H0Z"
+      android:fillColor="#00BBFF"/>
+</vector>
