+++ feature/search/src/main/java/com/kolown/search/component/DetailTopAppBar.kt
@@ -6,6 +6,8 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Close
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
@@ -13,29 +15,48 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import com.kolown.search.R
 
 @Composable
-fun DetailTopAppBar(
+internal fun DetailTopAppBar(
+    isReelsMode: Boolean,
+    onChangeReelsMode: (Boolean) -> Unit,
     popBackStack: () -> Unit,
 ) {
     Row(
         modifier = Modifier.fillMaxWidth().height(64.dp).background(Color.Transparent),
         verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.Start)
+        horizontalArrangement = if (isReelsMode) {
+            Arrangement.spacedBy(8.dp, Alignment.Start)
+        } else {
+            Arrangement.spacedBy(8.dp, Alignment.End)
+        }
     ) {
-        IconButton(
-            onClick = { popBackStack() }, modifier = Modifier.size(48.dp)
-        ) {
-            Icon(
-                imageVector = ImageVector.vectorResource(R.drawable.icon_arrow_back),
-                contentDescription = "뒤로 가기",
-                tint = Color.White,
-                modifier = Modifier.size(24.dp)
-            )
+        if (isReelsMode) {
+            IconButton(
+                onClick = { popBackStack() }, modifier = Modifier.size(48.dp)
+            ) {
+                Icon(
+                    imageVector = ImageVector.vectorResource(R.drawable.icon_arrow_back),
+                    contentDescription = "뒤로가기 버튼",
+                    tint = Color.White,
+                    modifier = Modifier.size(24.dp)
+                )
+            }
+        } else {
+            IconButton(
+                onClick = { onChangeReelsMode(true) },
+                modifier = Modifier.size(48.dp)
+            ) {
+                Icon(
+                    imageVector = Icons.Default.Close,
+                    contentDescription = "집중모드 해제 버튼",
+                    tint = Color.White,
+                    modifier = Modifier.size(24.dp)
+                )
+            }
         }
     }
 }
