+++ feature/search/src/main/java/com/kolown/search/navigation/SearchNavigation.kt
@@ -25,6 +25,8 @@ fun NavController.navigateSearchDetail() {
 fun NavGraphBuilder.searchNavGraph(
     padding: PaddingValues,
     navigateToTheir: (String) -> Unit,
+    isLoggedIn: Boolean,
+    onShowLoginSnackBar: () -> Unit,
     navigateToSearchDetail: () -> Unit,
     popBackStack: () -> Unit,
     getBackStackEntry: () -> NavBackStackEntry
@@ -48,7 +50,9 @@ fun NavGraphBuilder.searchNavGraph(
             padding = padding,
             viewModel = hiltViewModel(parentEntry),
             navigateToTheir = navigateToTheir,
-            popBackStack = popBackStack
+            popBackStack = popBackStack,
+            isLoggedIn = isLoggedIn,
+            onShowLoginSnackBar = onShowLoginSnackBar
         )
     }
 }
